+++ tests/test_ufrm.py
@@ -64,7 +64,8 @@ class UFRMTests(unittest.TestCase):
 
         # Check that all four expected fields are there.
         self.assertEqual(
-            set(('aff_bld', 'serv_bld', 'rnf_rt_idx', 'rnf_rt_m3')),
+            set(('aff_bld', 'serv_bld', 'rnf_rt_idx', 'rnf_rt_m3',
+                 'flood_vol')),
             set(field.GetName() for field in result_layer.schema))
 
         result_feature = next(result_layer)
@@ -72,7 +73,8 @@ class UFRMTests(unittest.TestCase):
                 ('aff_bld', 187010830.32202843),
                 ('serv_bld', 13253546667257.65),
                 ('rnf_rt_idx', 0.70387527942),
-                ('rnf_rt_m3', 70870.4765625)):
+                ('rnf_rt_m3', 70870.4765625),
+                ('flood_vol', 70870.484375)):
             result_val = result_feature.GetField(fieldname)
             places_to_round = (
                 int(round(numpy.log(expected_value)/numpy.log(10)))-6)